requirements.txt
py-cord-dev==2.5.0rc5
python-dotenv==1.0.0
psycopg2==2.9.9